+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 1.51 (not yet released)
 ------------
 
 * Added Formal Title to the action Contact: Create/Update Individual
+* Added action Contact: Create Note
 
 Version 1.50
 ------------